refactor: move WorkerEncryptionProxyProcessor
